requirements.txt
numpy~=1.19.2
Flask~=1.1.2
python-dotenv~=0.15.0